+++ lib/polkawallet_plugin_kusama.dart
@@ -34,6 +34,7 @@ import 'package:polkawallet_plugin_kusama/pages/staking/actions/stakePage.dart';
 import 'package:polkawallet_plugin_kusama/pages/staking/actions/stakingDetailPage.dart';
 import 'package:polkawallet_plugin_kusama/pages/staking/actions/unbondPage.dart';
 import 'package:polkawallet_plugin_kusama/pages/staking/validators/nominatePage.dart';
+import 'package:polkawallet_plugin_kusama/pages/staking/validators/validatorChartsPage.dart';
 import 'package:polkawallet_plugin_kusama/pages/staking/validators/validatorDetailPage.dart';
 import 'package:polkawallet_plugin_kusama/service/index.dart';
 import 'package:polkawallet_plugin_kusama/store/cache/storeCache.dart';
@@ -142,6 +143,7 @@ class PluginKusama extends PolkawalletPlugin {
       StakingDetailPage.route: (_) => StakingDetailPage(this, keyring),
       RewardDetailPage.route: (_) => RewardDetailPage(this, keyring),
       ValidatorDetailPage.route: (_) => ValidatorDetailPage(this, keyring),
+      ValidatorChartsPage.route: (_) => ValidatorChartsPage(this, keyring),
 
       // governance pages
       DemocracyPage.route: (_) => DemocracyPage(this, keyring),