+++ lib/service/staking.dart
@@ -105,17 +105,11 @@ class ApiStaking {
   }
 
   Future<Map> queryValidatorRewards(String accountId) async {
-    int timestamp = DateTime.now().second;
-    Map cached = store.staking.rewardsChartDataCache[accountId];
-    if (cached != null && cached['timestamp'] > timestamp - 1800) {
-      return cached;
-    }
     print('fetching rewards chart data');
     Map data = await api.staking.loadValidatorRewardsData(accountId);
     if (data != null) {
       // format rewards data & set cache
       Map chartData = PluginFmt.formatRewardsChartData(data);
-      chartData['timestamp'] = timestamp;
       store.staking.setRewardsChartData(accountId, chartData);
     }
     return data;